simulator/ethereum/engine: Refactor helper, add block modifiers
