+++ simulators/ethereum/engine/helper/payload.go
@@ -1,0 +1,313 @@
+package helper
+
+import (
+	"fmt"
+	"math/big"
+	"math/rand"
+	"strings"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
+	api "github.com/ethereum/go-ethereum/core/beacon"
+	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/hive/simulators/ethereum/engine/globals"
+)
+
+type CustomPayloadData struct {
+	ParentHash    *common.Hash
+	FeeRecipient  *common.Address
+	StateRoot     *common.Hash
+	ReceiptsRoot  *common.Hash
+	LogsBloom     *[]byte
+	PrevRandao    *common.Hash
+	Number        *uint64
+	GasLimit      *uint64
+	GasUsed       *uint64
+	Timestamp     *uint64
+	ExtraData     *[]byte
+	BaseFeePerGas *big.Int
+	BlockHash     *common.Hash
+	Transactions  *[][]byte
+}
+
+// Construct a customized payload by taking an existing payload as base and mixing it CustomPayloadData
+// BlockHash is calculated automatically.
+func CustomizePayload(basePayload *api.ExecutableDataV1, customData *CustomPayloadData) (*api.ExecutableDataV1, error) {
+	txs := basePayload.Transactions
+	if customData.Transactions != nil {
+		txs = *customData.Transactions
+	}
+	txsHash, err := calcTxsHash(txs)
+	if err != nil {
+		return nil, err
+	}
+	fmt.Printf("txsHash: %v\n", txsHash)
+	// Start by filling the header with the basePayload information
+	customPayloadHeader := types.Header{
+		ParentHash:  basePayload.ParentHash,
+		UncleHash:   types.EmptyUncleHash, // Could be overwritten
+		Coinbase:    basePayload.FeeRecipient,
+		Root:        basePayload.StateRoot,
+		TxHash:      txsHash,
+		ReceiptHash: basePayload.ReceiptsRoot,
+		Bloom:       types.BytesToBloom(basePayload.LogsBloom),
+		Difficulty:  big.NewInt(0), // could be overwritten
+		Number:      big.NewInt(int64(basePayload.Number)),
+		GasLimit:    basePayload.GasLimit,
+		GasUsed:     basePayload.GasUsed,
+		Time:        basePayload.Timestamp,
+		Extra:       basePayload.ExtraData,
+		MixDigest:   basePayload.Random,
+		Nonce:       types.BlockNonce{0}, // could be overwritten
+		BaseFee:     basePayload.BaseFeePerGas,
+	}
+
+	// Overwrite custom information
+	if customData.ParentHash != nil {
+		customPayloadHeader.ParentHash = *customData.ParentHash
+	}
+	if customData.FeeRecipient != nil {
+		customPayloadHeader.Coinbase = *customData.FeeRecipient
+	}
+	if customData.StateRoot != nil {
+		customPayloadHeader.Root = *customData.StateRoot
+	}
+	if customData.ReceiptsRoot != nil {
+		customPayloadHeader.ReceiptHash = *customData.ReceiptsRoot
+	}
+	if customData.LogsBloom != nil {
+		customPayloadHeader.Bloom = types.BytesToBloom(*customData.LogsBloom)
+	}
+	if customData.PrevRandao != nil {
+		customPayloadHeader.MixDigest = *customData.PrevRandao
+	}
+	if customData.Number != nil {
+		customPayloadHeader.Number = big.NewInt(int64(*customData.Number))
+	}
+	if customData.GasLimit != nil {
+		customPayloadHeader.GasLimit = *customData.GasLimit
+	}
+	if customData.GasUsed != nil {
+		customPayloadHeader.GasUsed = *customData.GasUsed
+	}
+	if customData.Timestamp != nil {
+		customPayloadHeader.Time = *customData.Timestamp
+	}
+	if customData.ExtraData != nil {
+		customPayloadHeader.Extra = *customData.ExtraData
+	}
+	if customData.BaseFeePerGas != nil {
+		customPayloadHeader.BaseFee = customData.BaseFeePerGas
+	}
+
+	// Return the new payload
+	return &api.ExecutableDataV1{
+		ParentHash:    customPayloadHeader.ParentHash,
+		FeeRecipient:  customPayloadHeader.Coinbase,
+		StateRoot:     customPayloadHeader.Root,
+		ReceiptsRoot:  customPayloadHeader.ReceiptHash,
+		LogsBloom:     customPayloadHeader.Bloom[:],
+		Random:        customPayloadHeader.MixDigest,
+		Number:        customPayloadHeader.Number.Uint64(),
+		GasLimit:      customPayloadHeader.GasLimit,
+		GasUsed:       customPayloadHeader.GasUsed,
+		Timestamp:     customPayloadHeader.Time,
+		ExtraData:     customPayloadHeader.Extra,
+		BaseFeePerGas: customPayloadHeader.BaseFee,
+		BlockHash:     customPayloadHeader.Hash(),
+		Transactions:  txs,
+	}, nil
+}
+
+func (customData *CustomPayloadData) String() string {
+	customFieldsList := make([]string, 0)
+	if customData.ParentHash != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("ParentHash=%s", customData.ParentHash.String()))
+	}
+	if customData.FeeRecipient != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("Coinbase=%s", customData.FeeRecipient.String()))
+	}
+	if customData.StateRoot != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("StateRoot=%s", customData.StateRoot.String()))
+	}
+	if customData.ReceiptsRoot != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("ReceiptsRoot=%s", customData.ReceiptsRoot.String()))
+	}
+	if customData.LogsBloom != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("LogsBloom=%v", types.BytesToBloom(*customData.LogsBloom)))
+	}
+	if customData.PrevRandao != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("PrevRandao=%s", customData.PrevRandao.String()))
+	}
+	if customData.Number != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("Number=%d", *customData.Number))
+	}
+	if customData.GasLimit != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("GasLimit=%d", *customData.GasLimit))
+	}
+	if customData.GasUsed != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("GasUsed=%d", *customData.GasUsed))
+	}
+	if customData.Timestamp != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("Timestamp=%d", *customData.Timestamp))
+	}
+	if customData.ExtraData != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("ExtraData=%v", *customData.ExtraData))
+	}
+	if customData.BaseFeePerGas != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("BaseFeePerGas=%s", customData.BaseFeePerGas.String()))
+	}
+	if customData.Transactions != nil {
+		customFieldsList = append(customFieldsList, fmt.Sprintf("Transactions=%v", customData.Transactions))
+	}
+	return strings.Join(customFieldsList, ", ")
+}
+
+type InvalidPayloadField string
+
+const (
+	InvalidParentHash             InvalidPayloadField = "ParentHash"
+	InvalidStateRoot                                  = "StateRoot"
+	InvalidReceiptsRoot                               = "ReceiptsRoot"
+	InvalidNumber                                     = "Number"
+	InvalidGasLimit                                   = "GasLimit"
+	InvalidGasUsed                                    = "GasUsed"
+	InvalidTimestamp                                  = "Timestamp"
+	InvalidPrevRandao                                 = "PrevRandao"
+	RemoveTransaction                                 = "Incomplete Transactions"
+	InvalidTransactionSignature                       = "Transaction Signature"
+	InvalidTransactionNonce                           = "Transaction Nonce"
+	InvalidTransactionGas                             = "Transaction Gas"
+	InvalidTransactionGasPrice                        = "Transaction GasPrice"
+	InvalidTransactionGasTipPrice                     = "Transaction GasTipCapPrice"
+	InvalidTransactionValue                           = "Transaction Value"
+	InvalidTransactionChainID                         = "Transaction ChainID"
+)
+
+// This function generates an invalid payload by taking a base payload and modifying the specified field such that it ends up being invalid.
+// One small consideration is that the payload needs to contain transactions and specially transactions using the PREVRANDAO opcode for all the fields to be compatible with this function.
+func GenerateInvalidPayload(basePayload *api.ExecutableDataV1, payloadField InvalidPayloadField) (*api.ExecutableDataV1, error) {
+
+	var customPayloadMod *CustomPayloadData
+	switch payloadField {
+	case InvalidParentHash:
+		modParentHash := basePayload.ParentHash
+		modParentHash[common.HashLength-1] = byte(255 - modParentHash[common.HashLength-1])
+		customPayloadMod = &CustomPayloadData{
+			ParentHash: &modParentHash,
+		}
+	case InvalidStateRoot:
+		modStateRoot := basePayload.StateRoot
+		modStateRoot[common.HashLength-1] = byte(255 - modStateRoot[common.HashLength-1])
+		customPayloadMod = &CustomPayloadData{
+			StateRoot: &modStateRoot,
+		}
+	case InvalidReceiptsRoot:
+		modReceiptsRoot := basePayload.ReceiptsRoot
+		modReceiptsRoot[common.HashLength-1] = byte(255 - modReceiptsRoot[common.HashLength-1])
+		customPayloadMod = &CustomPayloadData{
+			ReceiptsRoot: &modReceiptsRoot,
+		}
+	case InvalidNumber:
+		modNumber := basePayload.Number - 1
+		customPayloadMod = &CustomPayloadData{
+			Number: &modNumber,
+		}
+	case InvalidGasLimit:
+		modGasLimit := basePayload.GasLimit * 2
+		customPayloadMod = &CustomPayloadData{
+			GasLimit: &modGasLimit,
+		}
+	case InvalidGasUsed:
+		modGasUsed := basePayload.GasUsed - 1
+		customPayloadMod = &CustomPayloadData{
+			GasUsed: &modGasUsed,
+		}
+	case InvalidTimestamp:
+		modTimestamp := basePayload.Timestamp - 1
+		customPayloadMod = &CustomPayloadData{
+			Timestamp: &modTimestamp,
+		}
+	case InvalidPrevRandao:
+		// This option potentially requires a transaction that uses the PREVRANDAO opcode.
+		// Otherwise the payload will still be valid.
+		modPrevRandao := common.Hash{}
+		rand.Read(modPrevRandao[:])
+		customPayloadMod = &CustomPayloadData{
+			PrevRandao: &modPrevRandao,
+		}
+	case RemoveTransaction:
+		emptyTxs := make([][]byte, 0)
+		customPayloadMod = &CustomPayloadData{
+			Transactions: &emptyTxs,
+		}
+	case InvalidTransactionSignature,
+		InvalidTransactionNonce,
+		InvalidTransactionGas,
+		InvalidTransactionGasPrice,
+		InvalidTransactionGasTipPrice,
+		InvalidTransactionValue,
+		InvalidTransactionChainID:
+
+		if len(basePayload.Transactions) == 0 {
+			return nil, fmt.Errorf("No transactions available for modification")
+		}
+		var baseTx types.Transaction
+		if err := baseTx.UnmarshalBinary(basePayload.Transactions[0]); err != nil {
+			return nil, err
+		}
+		var customTxData CustomTransactionData
+		switch payloadField {
+		case InvalidTransactionSignature:
+			modifiedSignature := SignatureValuesFromRaw(baseTx.RawSignatureValues())
+			modifiedSignature.R = modifiedSignature.R.Sub(modifiedSignature.R, common.Big1)
+			customTxData.Signature = &modifiedSignature
+		case InvalidTransactionNonce:
+			customNonce := baseTx.Nonce() - 1
+			customTxData.Nonce = &customNonce
+		case InvalidTransactionGas:
+			customGas := uint64(0)
+			customTxData.Gas = &customGas
+		case InvalidTransactionGasPrice:
+			customTxData.GasPriceOrGasFeeCap = common.Big0
+		case InvalidTransactionGasTipPrice:
+			invalidGasTip := new(big.Int).Set(globals.GasTipPrice)
+			invalidGasTip.Mul(invalidGasTip, big.NewInt(2))
+			customTxData.GasTipCap = invalidGasTip
+		case InvalidTransactionValue:
+			// Vault account initially has 0x123450000000000000000, so this value should overflow
+			customTxData.Value, _ = hexutil.DecodeBig("0x123450000000000000001")
+		case InvalidTransactionChainID:
+			customChainID := new(big.Int).Set(globals.ChainID)
+			customChainID.Add(customChainID, common.Big1)
+			customTxData.ChainID = customChainID
+		}
+
+		modifiedTx, err := customizeTransaction(&baseTx, globals.VaultKey, &customTxData)
+		if err != nil {
+			return nil, err
+		}
+
+		modifiedTxBytes, err := modifiedTx.MarshalBinary()
+		if err != nil {
+			return nil, err
+		}
+		modifiedTransactions := [][]byte{
+			modifiedTxBytes,
+		}
+		customPayloadMod = &CustomPayloadData{
+			Transactions: &modifiedTransactions,
+		}
+	}
+
+	if customPayloadMod == nil {
+		return nil, fmt.Errorf("Invalid payload field to corrupt: %s", payloadField)
+	}
+
+	alteredPayload, err := CustomizePayload(basePayload, customPayloadMod)
+	if err != nil {
+		return nil, err
+	}
+
+	return alteredPayload, nil
+}
